requirements.txt
dbt-fabric
apache-airflow